added ntrain notebook, rename to OmniDeepSet
