+++ enreg/config/models/models.yaml
@@ -4,4 +4,4 @@ defaults:
   - HPS: HPS
   - SimpleDNN: SimpleDNN
   - OmniParT: OmniParT
-  - OmniFeedforward: OmniFeedforward
+  - OmniDeepSet: OmniDeepSet
